+++ R/cassette_class.R
@@ -165,7 +165,7 @@ Cassette <- R6::R6Class(
     #' @return character
     file = function() self$serializer$path,
 
-    #' @description is the cassette in recording mode?
+    #' @description Is the cassette in recording mode?
     #' @return logical
     recording = function() {
       if (self$record == "none") {
@@ -175,6 +175,12 @@ Cassette <- R6::R6Class(
       } else {
         TRUE
       }
+    },
+
+    #' @description Is the cassette in replaying mode?
+    #' @return logical
+    replaying = function() {
+      self$http_interactions$n_replayable() > 0
     },
 
     #' @description Remove outdated interactions